requirements.txt
torch>=2.3.0
torchaudio
accelerate
datasets
transformers
deepspeed
einops
wandb
npy-append-array
faiss-cpu
git+https://github.com/descriptinc/audiotools
